Add the abstract web initializer class for Spring Web MVC application
